% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors_matrix_fn.R
\name{create.W.neighbors.matrix.spConjNNGP}
\alias{create.W.neighbors.matrix.spConjNNGP}
\title{=============================================================================}
\usage{
create.W.neighbors.matrix.spConjNNGP(matrix.locations, n.neighbors)
}
\arguments{
\item{matrix.locations}{(matrix) observed locations (n.obs X 2)}

\item{n.neighbors}{(integer) number of neighbors to include in N(s)}
}
\value{
(list) the function returns the output of the spConjNNGP model
}
\description{
NEIGHBOR MATRICES CREATION FUNCTIONS
author: IG
date: 04-20-21
}
\details{
=============================================================================
neighbor matrix for osberved locations using spConjNNGP

creates matrix of neighbors for observations. The neighborhood of a location
"s" must only contain locations with smaller indexes. That is, the order of
the rows of the data matrix matters.
}
