% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors_matrix_fn.R
\name{create.W.neighbors.predictive.matrix}
\alias{create.W.neighbors.predictive.matrix}
\title{neighbor matrix for predicted locations}
\usage{
create.W.neighbors.predictive.matrix(
  Y,
  observed.locations,
  predicted.locations,
  n.neighbors
)
}
\arguments{
\item{n.neighbors}{(integer) number of neighbors to include in N(s)}

\item{matrix.distance}{(matrix) observed distance between the prediction and observed
locations (n.pred X n.obs)}

\item{fixed.neighbors}{(logical) set of 5 fixed neighbors for all locations
(default: FALSE)}
}
\value{
(matrix) the function returns a matrix (n.pred X n.obs) of logical to
indicate which observed location is a neighbor for each predicted location
}
\description{
creates matrix with observed neighbors index for each prediction location
}
