% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{marginal.likelihood.optim}
\alias{marginal.likelihood.optim}
\title{=============================================================================}
\usage{
marginal.likelihood.optim(
  pars,
  Y,
  D,
  Y_post,
  smallest.distance,
  W,
  n.obs,
  m,
  kappa,
  prior.list = list(a_sig = 3, b_sig = 2, a_tau = 3, b_tau = 2)
)
}
\arguments{
\item{pars}{(vector) of length four with the parameter values in the following
order: alpha, sigma, phi, xi}

\item{Y}{(vector) response value for our dataset}

\item{D}{(array) observed distance between each locations and its neighbors
(n.neighbors + 1 X n.neighbors + 1 X n.obs)}

\item{smallest.distance}{(numeric) approximate smallest observed distance between two
locations. This is used as a hyperparameter for the prior on the range parameter}

\item{W}{(matrix) matrix where each row s identifies the members of N(s)}

\item{n.obs}{(integer) number of observations in our dataset}

\item{m}{(integer) number of neighbors to include in N(s)}

\item{kappa}{(numeric) smoothness parameter for the Matern covariance function}

\item{Y.post}{(array) contains the matrix Y_{s, N(s)}'Y_{s, N(s)} for each s
(n.neighbors + 1 X n.neighbors + 1 X n.obs)}
}
\value{
(numeric) the function computes and returns the value of the
posterior marginal likelihood for a given set of parameters
}
\description{
OPTIMIZATION OF THE MARGINAL SPATIAL-T MODEL
author: IG
date: 04-28-21
}
\details{
=============================================================================
posterior marginal likelihood

compute the value of the posterior marginal likelihood
}
