% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{mv.marginal.likelihood.optim}
\alias{mv.marginal.likelihood.optim}
\title{mv posterior marginal likelihood}
\usage{
mv.marginal.likelihood.optim(
  pars,
  Y,
  D,
  Y_post,
  smallest.distance,
  W,
  Wnb,
  n.obs,
  m,
  kappa,
  cov.family,
  parsimonious,
  coregionalization
)
}
\arguments{
\item{pars}{(vector)}

\item{Y}{(vector) response value for our dataset}

\item{D}{(matrix) observed distance between all locations (n.obs X n.obs)}

\item{smallest.distance}{(numeric) smallest observed distance between two
locations in D}

\item{W}{(matrix) matrix where each row s identifies the members of N(s)}

\item{n.obs}{(integer) number of observations in our dataset}

\item{m}{(integer) number of neighbors to include in N(s)}

\item{kappa}{(numeric) smoothness parameter for the Matern covariance function}
}
\value{
(numeric) the function computes and returns the value of the
posterior marginal likelihood for a given set of parameters
}
\description{
compute the value of the bivariate posterior marginal likelihood
}
