% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_scoring_fn.R
\name{NIGP.coverage}
\alias{NIGP.coverage}
\title{Coverage scoring function}
\usage{
NIGP.coverage(Y, pp.samples)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{posterior.predictive.samples}{(matrix) posterior predictive samples
associated with the response vector Y. The number of rows must correspond
to the length of the vector Y.}
}
\value{
(numeric) the function returns the average coverage for the sample
}
\description{
This function computes the average coverage of the 95\% predictive intervals.
}
