% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_scoring_fn.R
\name{NIGP.crps}
\alias{NIGP.crps}
\title{=============================================================================}
\usage{
NIGP.crps(Y, posterior.predictive.samples)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{posterior.predictive.samples}{(matrix) posterior predictive samples
associated with the response vector Y. The number of rows must correspond
to the length of the vector Y.}
}
\value{
(numeric) the function returns the crps value
}
\description{
SCORING OF THE HIERARCHICAL SPATIAL-T MODEL
author: IG
date: 04-20-21
}
\details{
=============================================================================
CRPS scoring function

This function computes the continuous ranked probability score (CRPS) given
observations and posterior predictive samples.
}
