% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_scoring_fn.R
\name{NIGP.GG}
\alias{NIGP.GG}
\title{GG scoring function}
\usage{
NIGP.GG(Y, posterior.predictive.samples, k)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{posterior.predictive.samples}{(matrix) posterior predictive samples
associated with the response vector Y. The number of rows must correspond
to the length of the vector Y.}

\item{k}{(numeric) nonnegative penalty term from GG criterion}
}
\value{
(numeric) the function returns the GG value for the specified penalty k
}
\description{
This function computes the Gelfand & Ghosh (GG) score given a specific k,
observations and posterior predictive samples.
}
