% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_scoring_fn.R
\name{NIGP.pmse}
\alias{NIGP.pmse}
\title{PMSE scoring function}
\usage{
NIGP.pmse(Y, posterior.predictive.samples)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{posterior.predictive.samples}{(matrix) posterior predictive samples
associated with the response vector Y. The number of rows must correspond
to the length of the vector Y.}
}
\value{
(numeric) the function returns the pmse value
}
\description{
This function computes the predictive mean square error (PMSE) given
observations and posterior predictive samples.
}
