\name{NNRCM-package}
\alias{NNRCM-package}
\alias{NNRCM}
\docType{package}
\title{
\packageTitle{NNRCM}
}
\description{
\packageDescription{NNRCM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NNRCM}
\packageIndices{NNRCM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NNRCM}

Maintainer: \packageMaintainer{NNRCM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
<<<<<<< HEAD
>>>>>>> 3b936b19d828a1d798cf443b6d569bea5871d8c3:man/spatialT2-package.Rd
=======
>>>>>>> 72f76a1145d587c46514aa93607cdd96b17d39a9
