% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_inference_fn.R
\name{NNRCM.bv.hierarchical.infer}
\alias{NNRCM.bv.hierarchical.infer}
\title{Inference for the bivariate hierarchical spatial-t model}
\usage{
NNRCM.bv.hierarchical.infer(
  Y,
  observed.locations,
  X = 0,
  starting.values = list(alpha = 0, sigma = 0, nu = 0, nugget = 0),
  smoothness,
  a.nu.update.interval,
  priors = list(tau.prior = c(3, 2), a.prior = 1, nu.prior = c(5, 2), beta.prior = 0.1),
  tuning = list(a.tuning = 5, nu.tuning = 0.01),
  n.neighbors = 10,
  mcmc.samples = 1000,
  A = diag(2)
)
}
\arguments{
\item{Y}{}

\item{mcmc.samples}{}

\item{alpha}{}
}
\description{
the function uses MCMC to obtain posterior inference on the parameters of the
hierarchical model
}
