% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{NNRCM.bv.marginal.infer}
\alias{NNRCM.bv.marginal.infer}
\title{bivariate optim wrapper for marginal likelihood}
\usage{
NNRCM.bv.marginal.infer(
  Y,
  observed.locations,
  smoothness,
  starting.values = rep(0, 8),
  parsimonious = TRUE,
  mcmc.samples = 1000,
  n.neighbors = 10,
  coregionalization = FALSE,
  neighbor.info = FALSE,
  neighbor.seq = FALSE
)
}
\arguments{
\item{observed.locations}{(matrix) observed locations of bivariate data}

\item{smoothness}{(vector) vector of length three (v_{11}, v_r, v_{22}). The first and last terms are the
the Matern smoothness parameters of the responses. The middle term is the a number
between 1 and infinity which is used to compute v_{12}:
v_{12} = v_r / 2 * (v_{11} + v_{22})}

\item{starting.values}{(vector) starting values for the optimization, must be of
length four in the following order: alpha, sigma, phi, xi.}

\item{n.neighbors}{(integer) number of neighbors to include in N(s)}

\item{mydata}{(matrix) bivariate response value for our dataset. Use NA to represent missing data}

\item{cov.family}{(string) name of the covariance function (e.g. "matern")}
}
\value{
(list) the function returns a list with the output from the R::optim function
}
\description{
optimizes the posterior marginal likelihood of bivariate datasets
}
