% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{NNRCM.bv.marginal.predict}
\alias{NNRCM.bv.marginal.predict}
\title{posterior prediction function}
\usage{
NNRCM.bv.marginal.predict(
  Y,
  observed.locations,
  predicted.locations,
  point.estimates,
  n.sim = 1000,
  n.neighbors = 10,
  smoothness,
  coregionalization = FALSE,
  parsimonious = TRUE
)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{observed.locations}{(matrix) observed locations (n.obs X 2)}

\item{predicted.locations}{(matrix) predicted locations (n.pred X 2)}

\item{n.neighbors}{(integer) number of neighbors to include in N(s)}

\item{smoothness}{(numeric) smoothness parameter for the Matern covariance function}

\item{cov.pars}{(vector) vector of length two where the first element is the partial}

\item{cov.family}{(string) name of the covariance function (e.g. "matern")}

\item{alpha}{(numeric) degrees of freedom of the Wishart distribution}
}
\value{
(list) the function returns a list with the output from the R::optim function
}
\description{
predicts the response values for the requested set of predicted locations.
}
