% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_inference_fn.R
\name{NNRCM.hierarchical.infer}
\alias{NNRCM.hierarchical.infer}
\title{=============================================================================}
\usage{
NNRCM.hierarchical.infer(
  Y,
  observed.locations,
  X,
  starting.values,
  smoothness,
  nugget,
  a.nu.update.interval = 0,
  priors = list(tau.prior = c(3, 2), sig.prior = c(3, 2), a.prior = 1, nu.prior = c(5,
    2), beta.prior = 0.1),
  tuning = list(a.tuning = 5, nu.tuning = 0.01),
  n.neighbors = 10,
  mcmc.samples = 1000
)
}
\arguments{
\item{Y}{}

\item{mcmc.samples}{}

\item{alpha}{}
}
\description{
INFERENCE (MCMC) FOR THE HIERARCHICAL SPATIAL-T MODEL
author: IG
date: 04-20-21
}
\details{
=============================================================================
Inference for the hierarchical spatial-t model

the function uses MCMC to obtain posterior inference on the parameters of the
hierarchical model
}
