% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_prediction_fn.R
\name{NNRCM.hierarchical.predict}
\alias{NNRCM.hierarchical.predict}
\title{=============================================================================}
\usage{
NNRCM.hierarchical.predict(
  Y,
  X,
  observed.locations,
  predicted.locations,
  posterior.samples,
  point.estimates,
  smoothness,
  n.neighbors = 10
)
}
\description{
PREDICTION FOR THE HIERARCHICAL SPATIAL-T MODEL
author: IG
date: 04-20-21
}
\details{
=============================================================================
}
