% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{NNRCM.marginal.infer}
\alias{NNRCM.marginal.infer}
\title{optim wrapper for marginal likelihood}
\usage{
NNRCM.marginal.infer(
  Y,
  observed.locations,
  smoothness,
  n.neighbors = 10,
  starting.values = c(0, 0.5, 0.5, 0.5),
  prior.list = NULL
)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{observed.locations}{(matrix) observed locations for each observation
of mydata (n.obs X 2)}

\item{smoothness}{(numeric) smoothness parameter for the Matern covariance function}

\item{n.neighbors}{(integer) number of neighbors to include in N(s) (default: 10)}

\item{starting.values}{(vector) starting values for the optimization, must be of
length four in the following order: alpha, sigma, phi, xi
(default: (0, 0.5, 0.5, 0.5))}
}
\value{
(list) the function returns a list with the output from the R::optim
function
}
\description{
optimizes the posterior marginal likelihood
}
