% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_optim_fn.R
\name{NNRCM.marginal.predict}
\alias{NNRCM.marginal.predict}
\title{posterior prediction function}
\usage{
NNRCM.marginal.predict(
  Y,
  observed.locations,
  predicted.locations,
  optim.pars,
  smoothness,
  n.neighbors = 10,
  n.sim = 1000
)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{observed.locations}{(matrix) observed locations (n.obs X 2)}

\item{predicted.locations}{(matrix) predicted locations (n.pred X 2)}

\item{optim.pars}{(vector) vector of length four in the following order:
alpha, sigma, phi, tau}

\item{smoothness}{(numeric) smoothness parameter for the Matern covariance function}

\item{n.neighbors}{(integer) number of neighbors to include in N(s) (default: 10)}

\item{n.sim}{(integer) prediction sample size to output (default: 1000)}
}
\value{
(matrix) the function returns a matrix where each row is the prediction
samples for each predicted location.
}
\description{
predicts the response values for the requested set of predicted locations.
}
