% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_scoring_fn.R
\name{NNRCM.scores}
\alias{NNRCM.scores}
\title{Scoring function}
\usage{
NNRCM.scores(Y, pp.samples, k = 1)
}
\arguments{
\item{Y}{(vector) response value for our dataset}

\item{k}{(numeric) nonnegative penalty term from GG criterion}

\item{posterior.predictive.samples}{(matrix) posterior predictive samples
associated with the response vector Y. The number of rows must correspond
to the length of the vector Y.}
}
\value{
(numeric) the function returns the GG value for the specified penalty k
}
\description{
This function computes the Gelfand & Ghosh (GG) score, the continuous ranked
probability score (CRPS), the predictive mean squared error (PMSE) value
and the coverage given the true observations and the posterior predictive samples.
}
