% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sst}
\alias{sst}
\title{Sea Surface Temperature of Mediterranean Sea}
\format{
A data frame with 12211 rows and 3 variables:
\describe{
\item{longitude}{longitude of observation}
\item{latitude}{latitude of observation}
\item{sst_celsius}{sea surface temperature recored in celsius}
}
}
\usage{
data(sst)
}
\description{
A dataset containing the sea surface temperature recorded from
a variety of sources. The observations coming from buoys contain
many duplicates.
}
\keyword{datasets}
